cleanup(storage)!: relocate `ReadObjectResponse`
