+++ src/storage/examples/src/objects/stream_file_download.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 // [START storage_stream_file_download]
-use google_cloud_storage::ReadObjectResponse;
+use google_cloud_storage::read_object_response::ReadObjectResponse;
 use google_cloud_storage::client::Storage;
 
 pub async fn sample(client: &Storage, bucket_id: &str) -> anyhow::Result<()> {
