+++ src/storage/src/storage/read_object/resume_tests.rs
@@ -50,11 +50,11 @@
 
 use crate::model_ext::ReadRange;
 use crate::{
+    read_object_response::ReadObjectResponse,
     read_resume_policy::{ReadResumePolicyExt, Recommended},
     storage::client::tests::{
         MockBackoffPolicy, MockReadResumePolicy, MockRetryPolicy, MockRetryThrottler, test_builder,
     },
-    storage::read_object::ReadObjectResponse,
 };
 use gax::retry_policy::RetryPolicyExt;
 use gax::retry_result::RetryResult;
